Refactored for asynchronous localStorage

+++ frontend/src/redux/userLogin/userLoginActions.ts
@@ -3,7 +3,19 @@ import { LogInActionType } from './action-types';
 import { UserLogInActionInterface } from './userLoginInterfaces';
 
 import history from '../../historyObject';
-import authRequestAPI from '../../api/authRequest';
+import api from '../../api';
+
+//Async Await LocalStorage:
+const asyncLocalStorage = {
+    setItem: async function (key, value) {
+        await null;
+        return localStorage.setItem(key, value);
+    },
+    getItem: async function (key) {
+        await null;
+        return localStorage.getItem(key);
+    },
+};
 
 const userLogin =
     (formValues: {}) =>
@@ -14,7 +26,7 @@ const userLogin =
         let response;
 
         try {
-            response = await authRequestAPI.post('/users/login', {
+            response = await api.post('/users/login', {
                 ...formValues,
             });
         } catch (error) {
@@ -24,34 +36,36 @@ const userLogin =
             }
         }
 
-        // Store JWT in response into localstorage:
-        try {
-            //Outsourcing localstorage methods to promise to emulate asynchronous operation. When JWT is successfully set into localstorage, then we'll dispatch data to the state.
-            localStorage.setItem('jwt', JSON.stringify(response.data.token));
-
-            dispatch({
-                type: LogInActionType.USER_LOG_IN,
-                payload: response.data,
-            });
-
-            const jwtCheck = localStorage.getItem('jwt');
-
-            if (jwtCheck !== null && jwtCheck !== undefined) {
-                history.push('/dashboard');
-                //This is an incredibly inefficient and not react-minded compromise for the Bearer - Token issue. Hopefully in the future i'll think of something else.
-                window.location.reload();
-            }
-
-            // history.push('/dashboard');
-        } catch (e) {
-            if (e.name === 'QuotaExceededError') {
-                localStorage.clear();
-                localStorage.setItem('jwt', response.data.token);
-            } else {
-                alert('Your local storage seems to be full...');
-                console.log(e);
+        if (response) {
+            try {
+                asyncLocalStorage
+                    .setItem('jwt', JSON.stringify(response.data.token))
+                    .then(() => {
+                        return asyncLocalStorage
+                            .getItem('jwt')
+                            .then((value) => {
+                                if (value !== undefined && value !== null) {
+                                    history.push('/dashboard');
+                                }
+                            });
+                    });
+            } catch (e) {
+                if (e.name === 'QuotaExceededError') {
+                    localStorage.clear();
+                    localStorage.setItem('jwt', response.data.token);
+                } else {
+                    alert(
+                        `I'm Sorry! There seems to be a problem logging you in.`
+                    );
+                    console.log(e);
+                }
             }
         }
+
+        dispatch({
+            type: LogInActionType.USER_LOG_IN,
+            payload: response.data,
+        });
     };
 
 export default userLogin;
